+++ core/presentation/src/main/java/org/hesperides/core/presentation/io/platforms/properties/ValuedPropertyIO.java
@@ -65,28 +65,4 @@ public class ValuedPropertyIO extends AbstractValuedPropertyIO {
                 .map(ValuedPropertyIO::new)
                 .collect(Collectors.toSet());
     }
-
-    // `equals` et `hashCode` sont surchargés afin de reproduire le comportement du legacy qui est :
-    // une propriété valorisée doit être unique par son nom, lorsqu'on utilise un `Set`,
-    // c'est-à-dire lorsqu'on veut récupérer la liste des propriétés valorisées d'un module déployé
-    // => org.hesperides.core.presentation.io.platforms.properties.PropertiesIO
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(this.getName());
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        boolean equals;
-        if (this == object) {
-            equals = true;
-        } else if (object == null || getClass() != object.getClass()) {
-            equals = false;
-        } else {
-            final ValuedPropertyIO otherProperty = (ValuedPropertyIO) object;
-            equals = Objects.equals(this.getName(), otherProperty.getName());
-        }
-        return equals;
-    }
 }